WA-525 reduced file sizes and got rid of all but one warning
